When available, use /proc/$pid/map_files to read build-ids
This allows us to access binaries that are actually loaded into
a process. Reading build-ids from libraries and executables in the
filesystem is not reliable as those can be deleted or replaced
by different version by package update. When that happened,
rich-core-dumper reported wrong build-id for the binary.

Needs CONFIG_CHECKPOINT_RESTORE to be enabled in kernel.