Handle special case of first section's header in Quickie reports
Allow to omit initial newline of RICHCORE_HEADER at the beginning of
input file. Quick Feedback sends such reports and strictly speaking they
are malformed, but old rich-core-extract accepted the format and we
already have quite a number of those quickies collected, so let's
promote the short header to feature.